Create README.md
We believe in innovation, communication, and team work. We are a blockchain token that allows owners to share the total value of real estate owned, cash flow of rental properties, and profit from buying and selling of tax lien properties. We buy homes directly from home owners who are faced with losing their home to the tax lien auction.  Utilizing the BRRFR Method (Buy, Rehab, Rent, Fractionalize, Repeat) we are able to consistently grow the equity in our token for our holders. Once the home is rehabbed we will rent, NFT, and repeat the process. Our goal is to purchase as many tax lien homes yearly to return as much equity as possible to home owners while growing the equity backing the value of the Redeem Token.